requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
python-telegram-bot==21.3
resend==0.8.0
supabase==2.5.0
apscheduler==3.10.4
